Sim: Updated Propulsion Model
I updated the mini MUTT geometry and propulsion model. First the
geometry was originally defined in a coordinate system that was
different than the body fixed frame, which was error prone. So I defined
the c.g., aero center, and motor locations in a coordinate system
aligned with the body fixed frame and referenced at the nose of the
aircraft. Second, There was an error in the way the motor alignment was
defined (originally, this was put in as a yaw alignment, when it was
meant as a pitch alignment). Finally, I updated the locations and
alignments wrt data CR sent based on the SolidWorks model and validated
the forces and moments in the NL and SIL simulations.