FlightCode: Z dot controller
Created a new controller to enable z dot tracking and autolanding. Mode
0 enables the pilot to command theta and phi with an autothrottle
running. Mode 1 is meant for an initial approach, with a z dot tracker
running and autothrottle running, but the pilot commanding phi. Mode 2
is meant for short final / flare, with a z dot tracker running with a
lower reference command than the approach, an autothrottle running, and
a wing leveler running.