FlightCode: Updated Sys ID and CLAW
Updated sys id for the upcoming flight. 3 seconds waiting, then a
3-2-1-1 on surf pair 3, 8 seconds later a 3-2-1-1 on surf pair 4 and
repeat. Added a timer on the experimental mode to switch between flutter
supp on to off after 20 seconds into the experiment so we get both sets
of data. Added a timer to automate the throttle cutoff after speed drops
below 8 m/s in flare mode.