FlightCode: Memory Protection
The flight computers memory is limited to about 90000 data samples,
which was 30 minutes at the old rate of 50 Hz. Running at 150 Hz means
that this is now 10 minutes. In order to prevent flight data corruption
or loss, there is now protection where the system stops writing flight
data after 80000 samples (about 8 minutes). Operationally, this means
less of a rush getting to the aircraft and getting the data offloaded,
we'll just need to be careful to start the flight within ~3 minutes of
loading software.