FlightCode: Datalog trigger
Added a trigger to start the data logging. Now the datalog will start
after the pilot switches into autopilot for the first time. This means
for a flight op, shortly before flight, the pilot should switch into and
out of the autopilot mode to begin the datalog. After the initial
trigger, the log stays on until the memory protection kicks in about 8
minutes later. The upper left GCS light was used to illuminate when
logging starts.