FlightCode: Nav Filter Init
Modified the nav filter init to improve filter convergence. Pitch and
roll angle are initialized based on the accelerometer measurements (the
aircraft will need to be stationary during init, but this was always the
case due to grabbing sensor bias at that time). Heading is initialized
based off the magnetometer readings.