FlightCode: Datalog Bug Fix
Fixed a bug in datalog - the loop counter was moved in the code to
implement the trigger to begin recording. This move accidentally
introduced a bug where the data log loop counter, which is used to
protect the memory from over running, only ran in auto mode. This update
moves the loop counter and has been validated to fix this bug.