Flight Code: Throttle Mapping
Instead of auto endpoint detection on the ESC, we moved to fixed
endpoints, which were set using the DX9 transmitter. These endpoints
were set to PWM values of 1272 and 1792. The hati config was changed so
that the 0-1 throttle command will map now to 1272 and 1792 PWM values
instead of the 1000 to 2000 values used previously.